requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreschema==0.0.4
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
drf-yasg==1.21.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
django-modeltranslation==0.17.3